// types/pokemon-type.d.ts
declare interface PokemonType {
  type: {
    name:
      | 'fire'
      | 'water'
      | 'grass'
      | 'bug'
      | 'normal'
      | 'poison'
      | 'electric'
      | 'ground'
      | 'fairy'
      | 'fighting'
      | 'psychic'
      | 'rock'
      | 'ghost'
      | 'ice'
      | 'dragon'
      | 'dark'
      | 'steel'
      | 'flying';
  };
}

export default PokemonType;